+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - ensure all colums are for the same table in `ecto.ch.schema` mix task https://github.com/plausible/ecto_ch/pull/118
+- switch from `CREATE INDEX` to `ALTER TABLE ... ADD INDEX` syntax for indexes https://github.com/plausible/ecto_ch/pull/120
 
 ## 0.2.2 (2023-08-29)